تم تجهيز الاعدادات بالكامل مع الصلاحيت وتم تحديث جدول المستخدمين

+++ app/Dashboards/Settings/Index.php
@@ -32,26 +32,31 @@ class Index extends Component
 
     public function submit()
     {
-        $data = $this->validate(); 
-        if ($this->logo1) {
-            unlink('uplods/admin/settings/' .$this->settings->logo);
-            $imageName = 'logo' . '.' . $this->logo1->getClientOriginalExtension();
-            $this->logo1->storeAs('admin/settings/', $imageName, 'public');
-            $data['logo'] =  $imageName;
-        } else {
-            unset($data['logo']);
+        if (auth()->user()->hasPermission('edit_settings')) {
+            $data = $this->validate();
+            if ($this->logo1) {
+                if ($this->settings->logo) {
+                    unlink('uplods/admin/settings/' . $this->settings->logo);
+                }
+                $imageName = 'logo' . '.' . $this->logo1->getClientOriginalExtension();
+                $this->logo1->storeAs('admin/settings/', $imageName, 'public');
+                $data['logo'] =  $imageName;
+            } else {
+                unset($data['logo']);
+            }
+            if ($this->icon1) {
+                if ($this->settings->icon) {
+                    unlink('uplods/admin/settings/' . $this->settings->icon);
+                }
+                $imageName = 'icon' . '.' . $this->icon1->getClientOriginalExtension();
+                $this->icon1->storeAs('admin/settings/', $imageName, 'public');
+                $data['icon'] =  $imageName;
+            } else {
+                unset($data['icon']);
+            }
+            $this->settings->update($data);
+            session()->flash('message', 'Settings Updated Successfully');
         }
-        if ($this->icon1) {
-            unlink('uplods/admin/settings/' .$this->settings->icon);
-            $imageName = 'icon' . '.' . $this->icon1->getClientOriginalExtension();
-            $this->icon1->storeAs('admin/settings/', $imageName, 'public');
-            $data['icon'] =  $imageName;
-        } else {
-            unset($data['icon']);
-        }
-        // dd($data['settings.icon']);
-        $this->settings->update($data);
-        session()->flash('message', 'Settings Updated Successfully');
     }
 
     public function render()
